+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **Last and First Aggregation** when multiple values per day
+- **Empty Color** used when value is 0 can be configured
+- **Custom Color Theme** can be configured
 
 ### Changed